Update wg_manager.sh
Restructure code....bound to have introduced bugs!

Change Remove option '3+', instead the command now accepts parameters to eith list ALL ACTIVE Peers or a list of specific Peers. Also add 'full' keyword to show detailed report for the Peers)

Various other tweaks, such as if attempting to start a list of Peers and one of them is already ACTIVE, rather than abort the request with '***ERROR:', issue 'Warning:' and process the next Peer in the request.